requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
PyMySQL==0.9.3
# mysql-connector==2.2.9
Flask-SQLAlchemy==2.4.3
Flask-Migrate==2.5.3
Flask-Socketio
# Flask-Socketio==4.3.1
Flask-Mail==0.9.1
sqlalchemy==1.3.6
alembic==1.0.11
pytest==5.3.0
passlib==1.7.1
requests==2.22.0
# eventlet==0.29.1
gevent==20.9.0
Werkzeug==0.15.6
gunicorn==20.0.4
python-dotenv==0.13.0
redis==3.5.3
rq==1.6.1
reportlab==3.5.55
pdfkit==0.6.1
pyjwt==1.7.1
pyotp==2.4.1
stripe==2.55.1

# Full pip freeze
# alembic==1.0.11
# appdirs==1.4.4
# attrs==20.3.0
# black==20.8b1
# blinker==1.4
# certifi==2020.11.8
# chardet==3.0.4
# click==7.1.2
# dataclasses==0.8
# Flask==1.1.2
# Flask-Cors==3.0.9
# Flask-Mail==0.9.1
# Flask-Migrate==2.5.3
# Flask-SocketIO==4.3.1
# Flask-SQLAlchemy==2.4.3
# gevent==20.9.0
# greenlet==0.4.17
# gunicorn==20.0.4
# idna==2.8
# importlib-metadata==3.1.0
# itsdangerous==1.1.0
# Jinja2==2.11.2
# Mako==1.1.3
# MarkupSafe==1.1.1
# more-itertools==8.6.0
# mypy-extensions==0.4.3
# packaging==20.4
# passlib==1.7.1
# pathspec==0.8.1
# pdfkit==0.6.1
# Pillow==8.0.1
# pluggy==0.13.1
# py==1.9.0
# PyJWT==1.7.1
# PyMySQL==0.9.3
# pyotp==2.4.1
# pyparsing==2.4.7
# pytest==5.3.0
# python-dateutil==2.8.1
# python-dotenv==0.13.0
# python-editor==1.0.4
# python-engineio==3.13.2
# python-socketio==4.6.0
# redis==3.5.3
# regex==2020.11.13
# reportlab==3.5.55
# requests==2.22.0
# rq==1.6.1
# six==1.15.0
# SQLAlchemy==1.3.6
# stripe==2.55.1
# toml==0.10.2
# typed-ast==1.4.1
# typing-extensions==3.7.4.3
# urllib3==1.25.11
# wcwidth==0.2.5
# Werkzeug==0.15.6
# zipp==3.4.0
# zope.event==4.5.0
# zope.interface==5.2.0